+++ Package.swift
@@ -101,9 +101,7 @@ let package = Package(
         .target(
             name: "PayPalDataCollector",
             dependencies: ["BraintreeCore"],
-            path: "Sources/PayPalDataCollector",
-            exclude: ["Info.plist"],
-            publicHeadersPath: "Public"
+            path: "Sources/PayPalDataCollector"
         ),
         .binaryTarget(
             name: "PPRiskMagnes",