+++ Sources/BraintreeDataCollector/BTDataCollector_Internal.h
@@ -13,10 +13,5 @@
 */
 @property (nonatomic, strong, nonnull) KDataCollector *kount;
 
-/**
- The `PPDataCollector` class, exposed internally for injecting test doubles for unit tests
-*/
-+ (void)setPayPalDataCollectorClass:(nonnull Class)payPalDataCollectorClass;
-
 @end